Move data_model to external.
