+++ components/server/src/external/data_model/scales.py
