+++ components/server/src/external/data_model/sources/anchore.py
