+++ components/server/src/external/initialization/datamodel.py
@@ -5,8 +5,9 @@ import logging
 
 from pymongo.database import Database
 
-from data_model import DATA_MODEL_JSON
 from database.datamodels import insert_new_datamodel, latest_datamodel
+
+from ..data_model import DATA_MODEL_JSON
 
 
 def import_datamodel(database: Database) -> None:
