+++ components/server/tests/external/data_model/meta/test_metric.py
@@ -1,6 +1,6 @@
 """Unit tests for the metric meta model."""
 
-from data_model.meta.metric import Metric
+from external.data_model.meta.metric import Metric
 
 from .base import MetaModelTestCase
